predictive-maintenance/requirements.txt
fastapi==0.115.6
uvicorn==0.34.0
databricks-sdk>=0.38.0
mlflow>=2.19.0
psycopg2-binary==2.9.10
openai>=1.57.0
pydantic>=2.0
requests>=2.31.0
